aiplanner: WorldState / WorldFork split
Split WorldState into WorldState and WorldFork. WorldFork contains now all API functions that allow modification of the forked state. In a real life implementation we would not be able or want to modify the world state while we are planning, but simulate changes caused by individual actions. Therefore we allow the modification of a "fork" which has a local state overlay reflecting a hypothetical future.