genworldvoronoi: Profiling, settlement age, etc
Many different features that are in development:

Now humans spread from the cradle of civ across the continent and the sea, setting the age of the primary population centers.

Added population growth based on settlement age with some random disasters and stuff.

Started with the concept of history (non-functional ATM)